+++ public/javascript/qna.js
@@ -1,0 +1,16 @@
+//search event
+const transportSerachInfo = () => {
+  const inputValue = document.querySelector(".searchTerm").value;
+  location.href = `/qna?search=${inputValue}&list=1`;
+};
+const searchInput = document.querySelector(".searchTerm");
+searchInput.addEventListener("keyup", () => {
+  if (window.event.keyCode == 13) {
+    transportSerachInfo();
+  }
+});
+
+const searchBtn = document.querySelector(".searchButton");
+searchBtn.addEventListener("click", () => {
+  transportSerachInfo();
+});
