+++ public/stylesheets/qna.css
@@ -1,0 +1,111 @@
+/* qnaPage CSS */
+.findQna-wrap {
+  width: 100%;
+  height: 15em;
+  background-color: var(--yellow);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+.findQna-container {
+  display: flex;
+  width: 60%;
+  height: 90%;
+  justify-content: space-around;
+  align-items: center;
+}
+#findPeopleImg {
+  width: 15vw;
+  height: 19vh;
+  margin-right: 10vw;
+}
+
+.findText-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: flex-start;
+  width: 60%;
+  height: 70%;
+}
+.findText-container h2 {
+  color: white;
+  font-size: 1.8em;
+}
+.search {
+  width: 100%;
+  position: relative;
+  display: flex;
+}
+
+.searchTerm {
+  width: 70%;
+  border: 3px solid var(--bluemint);
+  border-right: none;
+  padding: 5px;
+  height: 36px;
+  border-radius: 5px 0 0 5px;
+  outline: none;
+  color: #9dbfaf;
+}
+
+.searchTerm:focus {
+  color: black;
+}
+
+.searchButton {
+  width: 40px;
+  height: 36px;
+  border: 1px solid var(--bluemint);
+  background: var(--bluemint);
+  text-align: center;
+  color: #fff;
+  border-radius: 0 5px 5px 0;
+  cursor: pointer;
+  font-size: 20px;
+}
+.search-wrap {
+  width: 100%;
+}
+.newQna-box {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+.newQna-box h3 {
+  color: white;
+  font-size: 1.5em;
+  margin-left: 1vw;
+}
+
+/* qnalist CSS */
+.qnaLists-wrap {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: flex-start;
+  margin-bottom: 1em;
+}
+.qnaLists-container {
+  width: 70%;
+  height: 80%;
+  padding-top: 5vh;
+}
+.qnaLists-container h1 {
+  padding-bottom: 3vh;
+  font-size: 2.5em;
+}
+.question-container {
+  display: flex;
+  padding: 0.6vw 0 0.6vw 0;
+  width: 100%;
+  border-bottom: 1px solid var(--darkgray);
+  font-family: "NanumSquareRound";
+  font-weight: bold;
+  font-size: 1.4em;
+  align-items: center;
+}
+.category {
+  margin-right: 1vw;
+  color: var(--yellow);
+}
