+++ views/include/_footer.ejs
@@ -3,7 +3,7 @@
       <div class="main-container">
         <div class="information-container">
           <div class="logo-container">
-            <img src="images/icon-logo-light.png" alt="">
+            <img src="/images/icon-logo-light.png" alt="">
             <a href="#">Schoolmate</a>
           </div>
           <div class="location-container">
@@ -11,11 +11,11 @@
             <p>경기도 안산시 단원구 와동 사세충열로 94</p>
           </div>
           <div class="mail-container">
-            <img src="images/icon-email-light.png" alt="" />
+            <img src="/images/icon-email-light.png" alt="" />
             <p>jungin7612@dimigo.hs.kr</p>
           </div>
           <div class="phone-container">
-            <img src="images/icon-phone-light.png" alt="">
+            <img src="/images/icon-phone-light.png" alt="">
             <p>010-2213-4083</p>
           </div>
         </div>
