+++ views/include/_header.ejs
@@ -11,7 +11,6 @@
       integrity="sha512-iBBXm8fW90+nuLcSKlbmrPcLa0OT92xO1BIsZ+ywDWZCvqsWgccV3gFoRBv0z+8dLJgyAHIhR35VZc2oM/gI1w=="
       crossorigin="anonymous"
     />
-    <script src="//code.jquery.com/jquery-3.4.1.min.js"></script>
   </head>
 
   <body>
@@ -22,13 +21,13 @@
           <p>경기도 안산시 단원구 와동 사세충열로 94</p>
         </div>
         <div class="mail-container">
-          <img src="images/icon-email.png" alt="" />
+          <img src="/images/icon-email.png" alt="" />
           <p>jungin7612@dimigo.hs.kr</p>
         </div>
       </nav>
       <div class="mainNav-container">
         <div class="logo-container">
-          <img src="images/icon-logo.png" alt="" />
+          <img src="/images/icon-logo.png" alt="" />
           <a href="/">Schoolmate</a>
         </div>
         <nav class="main-nav">
@@ -58,7 +57,7 @@
               <div class="profile-img"></div>
               <div class="welcome-container">
                 <p>안녕하세요.</p>
-                <p><span id="name"><%= name %> </span>학생</p>
+                <p><span id="name"><%= name %></span> 학생</p>
               </div>
               <i class="fas fa-bars aco-click"></i>
             </div>
