+++ views/index.ejs
@@ -1,6 +1,9 @@
 <!-- include header -->
 <%- include('include/_header') %>
 <!-- include header -->
+
+<link rel="stylesheet" href="/stylesheets/mainpage.css" />
+
 <div id="all-wrap">
   <div class="mainContent-wrap">
     <div class="mainContent-container">
