+++ views/qna.ejs
@@ -2,6 +2,8 @@
 <%- include('include/_header') %>
 <!-- include header -->
 
+<link rel="stylesheet" href="/stylesheets/qna.css" />
+
 <div class="findQna-Wrap">
   <div class="findQna-container">
     <img src="/images/qna-people.png" alt="" id="findPeopleImg" />
@@ -33,7 +35,7 @@
     <div class="paging"></div>
   </div>
 </div>
-
+<script src="/javascript/qna.js"></script>
 <!-- include footer -->
 <%- include('include/_footer') %>
 <!-- include footer -->
